tough: Always verify cached files before use
Every time a Repository is created, it should only trust the provided
trusted root metadata file, and not even its own cache. While we had
code to check for the cached root.json file, we should not trust such a
file if code consuming this library can't ensure a bad actor's inability
to touch the cache.

Don't check for cached root.json files, and always verify cached files
before we use them, ignoring them if they fail.

Because we trust the cached files before use, we don't need to check for
writable permissions on the cache.

Signed-off-by: iliana destroyer of worlds <iweller@amazon.com>