tough_schema: Rework trait bounds on Decoded<T>
None of the implementations for `Decoded<T>` actually require `T:
Decode` except for Deserialize, so let's drop that trait bound.

To make creating a `Decoded<T>` possible without using Deserialize,
there's now an Encode trait that is the reverse of Decode, and a
`From<Vec<u8>>` impl that uses it.

Signed-off-by: iliana destroyer of worlds <iweller@amazon.com>