Provide access to fetched signed metadata
This commit makes an API-breaking change to the `tough` library. It adds
`root`, `snapshot`, and `timestamp` methods to `Repository` which return
references to the signed metadata. The `targets` method for `Repository`
has been updated to do the same.

`DownloadArgs.run()` in `tuftool` has been updated to reflect the
aforementioned change to `Repository.targets()`.