Move signing functionality out of tuftool into tough
This commit makes an API-breaking change both to the `tough` library
and `tuftool`. It adds a new `Sign` trait to `tough`, which allows
users to implement this trait for different types of signing keys.
It also allows users to sign data using the `tough` library! :party:

A new method `canonical_form` is added to the `Role` trait, which
serializes the role into its canonical JSON form.

This commit removes the `KeyPair` enum and its corresponding methods
from `tuftool`. Areas that reference this enum have been updated to
use the new `Sign` trait.