Update targets methods in RepositoryEditor
This commit changes the `add_target` and `add_targets` methods
to `add_target_path` and `add_target_paths`. This is more descriptive
of their purpose. It also adds an `add_target` method, which takes
in a filename and a Target struct, which allows a user to parallelize
the creation of their targets.