Fix signature threshold
Prior to this PR, metadata signature verification would count multiple
signatures with identical authorized keyids each separately towards the
threshold.

This changes this to count unique valid keyids, and adds two tests to
check this behavior. One to simply copy the signature, and another to
re-sign the metadata twice with the non-deterministic rsassa-pss signing
scheme, to make sure we are not just deduplicating purely on the signature
value.