tough-kms: Fix handling of public key
The DER-encoded document we are getting from the GetPublicKey API is in
the SPKI format. `Decoded<RsaPem>`'s string form is SPKI, but its byte
form is RSAPublicKey.

Previously this code created the Decoded object from the byte form,
which caused it to be double-wrapped in an SPKI document.

Because we don't have access to tough's SPKI-parsing logic (not that any
other crates should), we encode the certificate as a PEM and create the
Decoded object from the string form.