tough: improve transport error interface
Make the inner error of the TransportError optional. Two public
functions are provided for constructing a TransportError. This is better
than a function that takes an Option<E> for the optional inner error
because the user would need to specify a type when trying to pass None.

Expand documentation of the TransportErrorKind::FileNotFound variant to
better explain the intent.

Hide the members of the TransportError struct to make breaking changes
less likely. Instead make them accessible by getter functions.

Changing the inner error to an Option made Snafu's macros less useful
because it gives special meaning to a struct member named 'source', and
would not compile when this became an Option. I wanted to keep the
member name because it seems to be canonical in Rust. Additionally, the
display macro became difficult to use to produce a clean error message.
It seemed easier to implement Error and Display myself.