tough: add tls features
Before this commit, there was no way to enable TLS for the HTTP
transport. The only way to do so was to manually include the same
version of reqwest and enable TLS on reqwest directly.

Here we keep the behavior we already had with the http feature (no TLS),
but we add two new features for the two TLS implementations.