speed up ci with caching
The commands we run during CI are broken into two Makefile targets:
- make build runs all the builds, unit tests and linting.
- make integ runs the http integration test using docker.

make ci is still available for local use, but in the GitHub actions
we run make build and make integ in parallel using a matrix.

We add a cache to both of the matrix jobs to speed them up. The cache is
keyed off of Cargo.lock, but includes a partial key restore so a cache
is still used even when Cargo.lock changes. It will then create a new
cache for future builds of the same Cargo.lock.

A file named cache_bust is added to the .github folder. Any edit to that
file will cause a complete cache miss. We can use this to start over if
we ever run into trouble.