tuftool: Add `download_root` module
This change adds a `download_root` module, which abstracts the logic of
downloading `root.json` out of `download` (the subcommand).  It also
makes the HTTP request a bit more robust by handling the initial request
error as well as a bad response.  Additionally, it avoids creating the
file until after the request has succeeded to avoid creating cruft on
the user's system.