tuftool: use `download_root` module in `download`
This change updates the `download` module/subcommand to make use of the
previously added `download_root` function.

It also defines a default of "1" for the `root_version` argument.
Previously, we effectively had this default in code by using
`1.root.json` in the event the argument wasn't passed.  It also has the
nice side effect of not needing to deal with an `Option` for this
argument.