tuftool: add `ignore-threshold` flag
With `18b7f740`, a check was added to verify that the root.json file was
signed with enough keys as specified in the thresholds. This is a
problem when a cloud provider is used to store keys, and they are in
different regions, since tuftool doesn't have the notion of regions
while it fetches the keys.

With the `ignore-threshold` flag, `tuftool` will be more "permissive"
and won't error out if a user attempts to sign the root.json file with
fewer keys than specified in the thresholds.

Signed-off-by: Arnaldo Garcia Rincon <agarrcia@amazon.com>