tough and tuftool: resolve target names
Adds a struct TargetName and replaces all uses of Strings as target
names so that we can resolve path segments such as ../

Adds structs for the two types of PathSet elements, PathPattern and
PathHashPrefix.

Improves the way that delegate target paths are checked.

Adds a function to the Repository object named save_target. This
function makes sure that a target can not be written outside of the
intended directory. Uses this function in tuftool

Tuftool now requires that the outdir not-exist for tuftool download and
uses the aforementioned save_target function for safer downloads.