Implement flexible parsing of ECDSA keys
TUF specification suggests to encode ECDSA keys using PEM. This is the
behaviour of the python reference implementation.

However, `go-tuf` encodes the keys as Hex numbers.

This commit makes the decoding of ECDSA keys more relaxed, the code will
try to decode them as PEM strings and it will resort to Hex when the
first decode fails.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>
Co-authored-by: Matthew James Briggs <matthew.james.briggs@gmail.com>