[RFC] add a lifetime parameter to Transport::fetch and Repository::read_target
This is useful in cases where you're implementing a transport that
borrows from the reader (e.g. if you're treating a zip file as the
reader, returning
[`ZipFile<'a>`](https://docs.rs/zip/latest/zip/read/struct.ZipFile.html)).

This is a breaking change because `read_target` now borrows from `self`.
I don't think that's a major issue, because in most cases you're going
to keep the `Repository` open for longer than any of the files.