tuftool: Allow specifying version in `root init`
A new repo is created by calling `tuftool root init <path>`. It is a
common pattern when renewing an expiring root to then have to call
`tuftool root bump-version` multiple times or `tuftool root
set-version`.

Since this is so common, this change makes it possible to provide an
option initial version to `root init` to avoid needing to run multiple
commmands. This adds an optional `--version` or `-v` argument that can
take a positive integer to set as the initial root version.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>