tuftool: Add transfer-metadata command
This adds a new `transfer-metadata` command to support migrating target
and metadata info to a new root. This would previously need to be done
by downloading all contents of a previous root and recreating and
recalculating SHAs for all targets. With many large targets, this
becomes an expensive operation. Since the previous root metadata already
contains this information, we can leverage that to just transfer the
metadata over to the new root.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>