tough: use default provider chains for KMS and SSM
Wrap the credential provider in a provider chain which adds an outer
refreshing cache provider. This is required for certain profiles to
work, for example a profile that uses `source_profile` to fetch role
credentials from a profile that uses `credential_process`, since the
presence of a cache is assumed.

Do the same for the region provider, as the region may come from the
environment or IMDS rather than being specified in the profile.

Signed-off-by: Ben Cressey <bcressey@amazon.com>