deps: Clean up base64 references
With the recent upgrades of pem and base64 being done in two different
PRs, there were a few leftover things to clean up related to base64.

The Cargo.lock file still has a reference to base64 0.13.1. With both
pem and base64 updated, there actually wasn't anything referencing this
version anymore and any references to 0.13.1 should be removed from the
lock file.

With the newer pem dependency, there is not a transient dependency to
the older base64 being pulled in. The cargo deny configuration still had
a reference to this older version that is no longer needed.

There have also been some minor changes in how encode/decode are called
with the base64 library. A deprecated method was still being used. That
is now updated to the preferred/non-deprecated method.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>