deny.toml: allow-wildcard-paths
failure-server is an internal package used only in testing that we do
not want to publish to crates.io. When including it in tough's dev-
dependencies, it is necessary to do so without a version number or else
cargo publish will fail. The lack of version number causes cargo deny
to ban failure-server as a wildcard-versioned dependency. The following
line allows us to refer to our own internal packages without a version
number.