Exclude check-license workflows for non-Linux
We now have a matrix job to ensure test coverage on all supported OS
platforms. This also causes the `check-license` job to run on each OS.

The license check isn't platform specific, so this ends up being a
little wasteful. This adds an exclude block to skip macOS and Windows
since they are not needed and those particular OS's take longer to run
than Linux.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>