tuftool: fix root gen-rsa command bug
A bug introduced in 55a40cc caused the tuftool tuftool root gen-rsa
command to be unusable because it caused file paths to be parsed with
Url::parse. This commit only uses URL for parsing if the SSM or KMS
schemes are found. Otherwise it is assumed to be a file path and parsed
with PathBuf.