github-actions: update Mac runners to macos-15
I've been observing some strange test failures on the Mac runners on
some of my recent PRs. The `tuftool clone` tests fail about 1 in every
10 runs, and every time it's because the `1.root.json` that was copied
is actually just an empty file.

This hasn't happened on runners for any other OS, and when I tried to
replicate it on my Mac that's running MacOS 15, I wasn't able to.
Attempting to bump the runners to macos-15 to see if it resolves this
issue, otherwise I will continue investigating this as a potential bug.

Signed-off-by: Sam Berning <bernings@amazon.com>